Refactor basic consensus message checking
